+++ src/mpc/structs.jl
@@ -226,7 +226,15 @@ Base.@kwdef struct MPCElectricStorage < AbstractElectricStorage
     soc_min_fraction::Float64 = 0.2
     soc_init_fraction::Float64 = 0.5
     can_grid_charge::Bool = true
+    can_net_meter::Bool = true
+    can_wholesale::Bool = false
+    can_export_beyond_nem_limit = false
     grid_charge_efficiency::Float64 = 0.96 * 0.975^2
+    max_kw::Float64 = size_kw
+    max_kwh::Float64 = size_kwh
+    minimum_avg_soc_fraction::Float64 = 0.0
+    fixed_soc_series_fraction::Union{Nothing, Array{<:Real,1}} = nothing
+    fixed_soc_series_fraction_tolerance::Real = !isnothing(fixed_soc_series_fraction) ? 0.1 : 0.0
 end
 ```
 """
@@ -238,10 +246,15 @@ Base.@kwdef struct MPCElectricStorage <: AbstractElectricStorage
     soc_min_fraction::Float64 = 0.2
     soc_init_fraction::Float64 = 0.5
     can_grid_charge::Bool = true
+    can_net_meter::Bool = true
+    can_wholesale::Bool = false
+    can_export_beyond_nem_limit = false
     grid_charge_efficiency::Float64 = 0.96 * 0.975^2
     max_kw::Float64 = size_kw
     max_kwh::Float64 = size_kwh
     minimum_avg_soc_fraction::Float64 = 0.0
+    fixed_soc_series_fraction::Union{Nothing, Array{<:Real,1}} = nothing
+    fixed_soc_series_fraction_tolerance::Real = !isnothing(fixed_soc_series_fraction) ? 0.1 : 0.0
 end